params/view type update
