+++ nexus/db-model/Cargo.toml
@@ -36,3 +36,7 @@ nexus-defaults.workspace = true
 nexus-types.workspace = true
 omicron-passwords.workspace = true
 sled-agent-client.workspace = true
+
+[dev-dependencies]
+expectorate.workspace = true
+strum.workspace = true
