+++ nexus/db-model/src/lib.rs
@@ -309,6 +309,76 @@ pub trait DatabaseString: Sized {
     fn from_database_string(s: &str) -> Result<Self, Self::Error>;
 }
 
+use anyhow::anyhow;
+use nexus_types::external_api::shared::FleetRole;
+use nexus_types::external_api::shared::ProjectRole;
+use nexus_types::external_api::shared::SiloRole;
+
+impl DatabaseString for FleetRole {
+    type Error = anyhow::Error;
+
+    fn to_database_string(&self) -> &str {
+        match self {
+            FleetRole::Admin => "admin",
+            FleetRole::Collaborator => "collaborator",
+            FleetRole::Viewer => "viewer",
+        }
+    }
+
+    fn from_database_string(s: &str) -> Result<Self, Self::Error> {
+        match s {
+            "admin" => Ok(FleetRole::Admin),
+            "collaborator" => Ok(FleetRole::Collaborator),
+            "viewer" => Ok(FleetRole::Viewer),
+            _ => Err(anyhow!("unsupported Fleet role from database: {:?}", s)),
+        }
+    }
+}
+
+impl DatabaseString for SiloRole {
+    type Error = anyhow::Error;
+
+    fn to_database_string(&self) -> &str {
+        match self {
+            SiloRole::Admin => "admin",
+            SiloRole::Collaborator => "collaborator",
+            SiloRole::Viewer => "viewer",
+        }
+    }
+
+    fn from_database_string(s: &str) -> Result<Self, Self::Error> {
+        match s {
+            "admin" => Ok(SiloRole::Admin),
+            "collaborator" => Ok(SiloRole::Collaborator),
+            "viewer" => Ok(SiloRole::Viewer),
+            _ => Err(anyhow!("unsupported Silo role from database: {:?}", s)),
+        }
+    }
+}
+
+impl DatabaseString for ProjectRole {
+    type Error = anyhow::Error;
+
+    fn to_database_string(&self) -> &str {
+        match self {
+            ProjectRole::Admin => "admin",
+            ProjectRole::Collaborator => "collaborator",
+            ProjectRole::Viewer => "viewer",
+        }
+    }
+
+    fn from_database_string(s: &str) -> Result<Self, Self::Error> {
+        match s {
+            "admin" => Ok(ProjectRole::Admin),
+            "collaborator" => Ok(ProjectRole::Collaborator),
+            "viewer" => Ok(ProjectRole::Viewer),
+            _ => {
+                Err(anyhow!("unsupported Project role from database: {:?}", s))
+            }
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::VpcSubnet;
@@ -431,5 +501,93 @@ mod tests {
         assert!(subnet.check_requestable_addr("fd00::a".parse().unwrap()));
         assert!(!subnet.check_requestable_addr("fd00::1".parse().unwrap()));
         assert!(subnet.check_requestable_addr("fd00::1:1".parse().unwrap()));
+    }
+
+    /// Does some basic smoke checks on an impl of `DatabaseString`
+    ///
+    /// This tests:
+    ///
+    /// - that for every variant, if we serialize it and deserialize the result,
+    ///   we get back the original variant
+    /// - that if we attempt to deserialize some _other_ input, we get back an
+    ///   error
+    /// - that the serialized form for each variant matches what's found in
+    ///   `expected_output_file`.  This output file is generated by this test
+    ///   using expectorate.
+    ///
+    /// This cannot completely test the correctness of the implementation, but it
+    /// can catch some basic copy/paste errors and accidental compatibility
+    /// breakage.
+    #[cfg(test)]
+    pub fn test_database_string_impl<T, P>(expected_output_file: P)
+    where
+        T: std::fmt::Debug
+            + PartialEq
+            + super::DatabaseString
+            + strum::IntoEnumIterator,
+        P: std::convert::AsRef<std::path::Path>,
+    {
+        let mut output = String::new();
+        let mut maxlen: Option<usize> = None;
+
+        for variant in T::iter() {
+            // Serialize the variant.  Verify that we can deserialize the thing
+            // we just got back.
+            let serialized = variant.to_database_string();
+            let deserialized = T::from_database_string(serialized)
+                .unwrap_or_else(|_| {
+                    panic!(
+                        "failed to deserialize the string {:?}, which we \
+                    got by serializing {:?}",
+                        serialized, variant
+                    )
+                });
+            assert_eq!(variant, deserialized);
+
+            // Put the serialized form into "output".  At the end, we'll compare
+            // this to the expected output.  This will fail if somebody has
+            // incompatibly changed things.
+            output.push_str(&format!(
+                "variant {:?}: serialized form = {}\n",
+                variant, serialized
+            ));
+
+            // Keep track of the maximum length of the serialized strings.
+            // We'll use this to construct an input to `from_database_string()`
+            // that was not emitted by any of the variants'
+            // `to_database_string()` functions.
+            match (maxlen, serialized.len()) {
+                (None, newlen) => maxlen = Some(newlen),
+                (Some(oldlen), newlen) if newlen > oldlen => {
+                    maxlen = Some(newlen)
+                }
+                _ => (),
+            }
+        }
+
+        // Check that `from_database_string()` fails when given input that
+        // doesn't match any of the variants' serialized forms.  We construct
+        // this input by providing a string that's longer than all strings
+        // emitted by `to_database_string()`.
+        if let Some(maxlen) = maxlen {
+            let input = String::from_utf8(vec![b'-'; maxlen + 1]).unwrap();
+            T::from_database_string(&input)
+                .expect_err("expected failure to deserialize unknown string");
+        }
+
+        expectorate::assert_contents(expected_output_file, &output);
+    }
+
+    #[test]
+    fn test_roles_database_strings() {
+        test_database_string_impl::<super::FleetRole, _>(
+            "tests/output/authz-roles-fleet.txt",
+        );
+        test_database_string_impl::<super::SiloRole, _>(
+            "tests/output/authz-roles-silo.txt",
+        );
+        test_database_string_impl::<super::ProjectRole, _>(
+            "tests/output/authz-roles-project.txt",
+        );
     }
 }
