+++ nexus/db-queries/Cargo.toml
@@ -36,7 +36,6 @@ openssl.workspace = true
 openssl-sys.workspace = true
 openssl-probe.workspace = true
 oso.workspace = true
-parse-display.workspace = true
 paste.workspace = true
 # See omicron-rpaths for more about the "pq-sys" dependency.
 pq-sys = "*"
@@ -45,7 +44,6 @@ ref-cast.workspace = true
 reqwest = { workspace = true, features = [ "json" ] }
 ring.workspace = true
 samael.workspace = true
-schemars = { workspace = true, features = ["chrono", "uuid1"] }
 serde.workspace = true
 serde_json.workspace = true
 serde_urlencoded.workspace = true
