+++ nexus/db-queries/src/db/fixed_data/silo.rs
@@ -22,6 +22,7 @@ lazy_static! {
             identity_mode: shared::SiloIdentityMode::LocalOnly,
             admin_group_name: None,
             tls_certificates: vec![],
+            mapped_fleet_roles: Default::default(),
         },
-    );
+    ).unwrap();
 }
