+++ nexus/test-utils/src/resource_helpers.rs
@@ -263,6 +263,7 @@ pub async fn create_silo(
             identity_mode,
             admin_group_name: None,
             tls_certificates: vec![],
+            mapped_fleet_roles: Default::default(),
         },
     )
     .await
