+++ nexus/types/Cargo.toml
@@ -13,10 +13,12 @@ newtype_derive.workspace = true
 openssl.workspace = true
 openssl-sys.workspace = true
 openssl-probe.workspace = true
+parse-display.workspace = true
 schemars = { workspace = true, features = ["chrono", "uuid1"] }
 serde.workspace = true
 serde_json.workspace = true
 steno.workspace = true
+strum.workspace = true
 uuid.workspace = true
 
 api_identity.workspace = true
