+++ nexus/types/src/external_api/views.rs
@@ -17,6 +17,7 @@ use omicron_common::api::external::{
 };
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
+use std::collections::BTreeMap;
 use std::net::IpAddr;
 use uuid::Uuid;
 
@@ -36,6 +37,13 @@ pub struct Silo {
 
     /// How users and groups are managed in this Silo
     pub identity_mode: shared::SiloIdentityMode,
+
+    /// Mapping of which Fleet roles are conferred by each Silo role
+    ///
+    /// The default is that no Fleet roles are conferred by any Silo roles
+    /// unless there's a corresponding entry in this map.
+    #[serde(default)]
+    pub mapped_fleet_roles: BTreeMap<shared::SiloRole, Vec<shared::FleetRole>>,
 }
 
 // IDENTITY PROVIDER
